requirements/base.txt
djangorestframework-simplejwt>=4.7,<4.8
